+++ EaSElect/Model/DB.Context.cs
@@ -54,15 +54,15 @@ namespace Models
             return ((IObjectContextAdapter)this).ObjectContext.ExecuteFunction<Nullable<int>>("IsExistTypeDetails", typeDetailNameParameter);
         }
     
-        public virtual ObjectResult<Nullable<int>> IsExistVoter(Nullable<int> voterId, Nullable<int> electionId)
+        public virtual ObjectResult<Nullable<int>> IsExistVoter(Nullable<long> voterId, Nullable<long> electionId)
         {
             var voterIdParameter = voterId.HasValue ?
                 new ObjectParameter("voterId", voterId) :
-                new ObjectParameter("voterId", typeof(int));
+                new ObjectParameter("voterId", typeof(long));
     
             var electionIdParameter = electionId.HasValue ?
                 new ObjectParameter("electionId", electionId) :
-                new ObjectParameter("electionId", typeof(int));
+                new ObjectParameter("electionId", typeof(long));
     
             return ((IObjectContextAdapter)this).ObjectContext.ExecuteFunction<Nullable<int>>("IsExistVoter", voterIdParameter, electionIdParameter);
         }
